fix: Avaliador

+++ src/controllers/rentals/deleteRentalsIDController.js
@@ -12,6 +12,8 @@ export async function DeleteRentalsID (req, res) {
 
         if(!results) return res.sendStatus(404);
 
+        if(results.returnDate !== null) return res.sendStatus(400);
+
         await db.query(`DELETE FROM rentals WHERE id=${id}`);
 
         res.sendStatus(200);
